backend/djreact/requirements.txt
asgiref==3.4.1
astroid==2.7.2
autopep8==1.5.7
Django==3.2.6
django-cors-headers==3.8.0
djangorestframework==3.12.4
flake8==3.9.2
isort==5.9.3
lazy-object-proxy==1.6.0
mccabe==0.6.1
mypy==0.910
mypy-extensions==0.4.3
platformdirs==2.2.0
pycodestyle==2.7.0
pyflakes==2.3.1
pylint==2.10.2
pylint-django==2.4.4
pylint-plugin-utils==0.6
pytz==2021.1
sqlparse==0.4.1
toml==0.10.2
typing-extensions==3.10.0.0
wrapt==1.12.1
